python/2_phonebot/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.37.1"
environs = "^11.0.0"
fastapi = {extras = ["standard"], version = "^0.112.0"}
retell-sdk = "^4.2.0"

[tool.poetry.group.lint.dependencies]
ruff = "^0.4.9"


[tool.poetry.group.debug.dependencies]
ipython = "^8.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
